+++ test/cacheTest.mjs
@@ -1,7 +1,7 @@
 import { test } from 'uvu';
 import * as assert from 'uvu/assert';
 import { build } from "esbuild";
-import { mkdirSync, writeFileSync } from "fs";
+import { mkdirSync, unlinkSync, writeFileSync } from "fs";
 import { sass } from "svelte-preprocess-sass";
 import sveltePlugin from "../dist/index.mjs";
 import { tmpdir } from 'os';
@@ -48,9 +48,10 @@ test("Cache w/ rebuild", async () => {
     await incrementalTest();
 });
 
-test("Preprocess w/ deps", (async () => {
+//set up a basic incremental build for use in other tests
+async function depsSetup() {
     const dirname = join(tmpdir(), "esbuild-svelte");
-    mkdirSync(dirname, {recursive: true});
+    mkdirSync(dirname, { recursive: true });
     writeFileSync(join(dirname, '/app.js'), 'import x from "./foo.svelte"\nconsole.log(x)');
     writeFileSync(join(dirname, '/foo.svelte'), '<style lang="sass">@import "./xyz.sass"</style><div class="xyz">foo</div>');
 
@@ -68,16 +69,66 @@ test("Preprocess w/ deps", (async () => {
                 style: sass(),
             },
         })],
-        logLevel: 'info',
+        logLevel: 'silent',
     });
-    assert.match(result.outputFiles[1].text, "red");
+    return { result, dirname };
+}
+
+test("Preprocess w/ deps basic", (async () => {
+    let { result, dirname } = await depsSetup();
 
     // Set color to green
     writeFileSync(join(dirname, '/xyz.sass'), '.xyz\n  color: green');
     const result2 = await result.rebuild();
-    assert.match(result2.outputFiles[1].text, "green");
 
     result.rebuild.dispose();
+
+
+    assert.match(result.outputFiles[1].text, "red");
+    assert.match(result2.outputFiles[1].text, "green");
+}));
+
+test("Preprocess w/ deps delete", (async () => {
+    let { result, dirname } = await depsSetup();
+
+    // remove file
+    unlinkSync(join(dirname, '/xyz.sass'));
+    try {
+        await result.rebuild();
+        assert.not.ok(true);
+    } catch (err) {
+        assert.ok(err.errors, "Rebuild should generate an exception when file deleted")
+        assert.not.match(err.errors[0].text, "stat", "`stat` shouldn't be generating the error message")
+        assert.ok(err.errors.length == 1, "There should be one error");
+    } finally {
+        result.rebuild.dispose();
+    }
+}));
+
+test("Don't cache errors", (async () => {
+    let { result, dirname } = await depsSetup();
+
+    // remove file
+    unlinkSync(join(dirname, '/xyz.sass'));
+    try {
+        await result.rebuild();
+        assert.not.ok(true);
+    } catch (err) {
+        assert.ok(err.errors);//expect to throw from rebuild
+    }
+    
+    // bring it back
+    // the previous error should go away
+    writeFileSync(join(dirname, '/xyz.sass'), '.xyz\n  color: green');
+    try {
+        await result.rebuild();
+        assert.ok(true);
+    } catch (err) {
+        assert.not.ok(true);//expect to not to
+    } 
+    finally {
+        result.rebuild.dispose();
+    }
 }));